Add Two Players Functionality

+++ static/engines/tetrisEngine.js
@@ -17,39 +17,41 @@ function createPiece(type) {
   }
 }
 
-function updateScore() {
-  document.getElementById("score").innerText = tetris.player.score;
-}
+const tetri = [];
 
-const canvas = document.getElementById("tetris");
-const tetris = new Tetris(canvas);
+const playerElements = document.querySelectorAll(".player");
+[...playerElements].forEach(element => {
+  const tetris = new Tetris(element);
+  tetri.push(tetris);
+});
 
-const handleKeyPress = event => {
-  const player = tetris.player;
-  switch (event.keyCode) {
-    case 37:
-      player.move(-1);
-      break;
-    case 38:
-      break;
-    case 39:
-      player.move(1);
-      break;
-    case 40:
-      player.drop();
-      break;
-    case 81:
-      player.rotate(-1);
-      break;
-    case 87:
-      player.rotate(1);
-      break;
-    default:
-      console.log("Other Key", event.key, "Key Code", event.keyCode);
-      break;
-  }
+const keyListener = event => {
+  [[65, 68, 81, 69, 83], [72, 75, 89, 73, 74]].forEach((key, index) => {
+    const player = tetri[index].player;
+    if (event.type === "keydown") {
+      if (event.keyCode === key[0]) {
+        player.move(-1);
+      } else if (event.keyCode === key[1]) {
+        player.move(1);
+      } else if (event.keyCode === key[2]) {
+        player.rotate(-1);
+      } else if (event.keyCode === key[3]) {
+        player.rotate(1);
+      }
+    }
+
+    if (event.keyCode === key[4]) {
+      if (event.type === "keydown") {
+        if (player.dropInterval !== player.DROP_FAST) {
+          player.drop();
+          player.dropInterval = player.DROP_FAST;
+        }
+      } else {
+        player.dropInterval = player.DROP_SLOW;
+      }
+    }
+  });
 };
 
-document.addEventListener("keydown", handleKeyPress);
-
-updateScore();
+document.addEventListener("keydown", keyListener);
+document.addEventListener("keyup", keyListener);
